- 지훈  객실, 부가서비스 요금 변경 오류 변수값 long 변경

+++ kr/or/bit/hotel/Hotel.java
@@ -13,10 +13,10 @@ public abstract class Hotel implements Serializable {
 	private Room[] roomInfos;
 	private long sales;
 	private Map<String, Member> members;
-	private int[] roomPrices;
-	private int breakfast;
-	private int therapy;
-	private int[] servicePrices;
+	private long[] roomPrices;
+	private long breakfast;
+	private long therapy;
+	private long[] servicePrices;
 	private LocalDate today;
 
 	public Hotel() {
@@ -26,7 +26,7 @@ public abstract class Hotel implements Serializable {
 		today = LocalDate.now();
 		breakfast = Number.breakfast;
 		therapy = Number.therapy;
-		servicePrices = new int[] { Number.breakfast, Number.therapy };
+		servicePrices = new long[] { Number.breakfast, Number.therapy };
 		setHotelRoomPrices();
 		setHotelRooms();
 		setHotelRoomNumber();
@@ -44,7 +44,7 @@ public abstract class Hotel implements Serializable {
 		}
 	}
 
-	public int getBreakfast() {
+	public long getBreakfast() {
 		return breakfast;
 	}
 
@@ -52,7 +52,7 @@ public abstract class Hotel implements Serializable {
 		this.breakfast = breakfast;
 	}
 
-	public int getTherapy() {
+	public long getTherapy() {
 		return therapy;
 	}
 
@@ -64,11 +64,11 @@ public abstract class Hotel implements Serializable {
 		this.rooms = rooms;
 	}
 
-	public void setRoomPrices(int[] roomPrices) {
+	public void setRoomPrices(long[] roomPrices) {
 		this.roomPrices = roomPrices;
 	}
 
-	public int[] getRoomPrices() {
+	public long[] getRoomPrices() {
 		return roomPrices;
 	}
 
@@ -100,11 +100,11 @@ public abstract class Hotel implements Serializable {
 		return rooms;
 	}
 
-	public int[] getServicePrices() {
+	public long[] getServicePrices() {
 		return servicePrices;
 	}
 
-	public void setServicePrices(int[] servicePrices) {
+	public void setServicePrices(long[] servicePrices) {
 		this.servicePrices = servicePrices;
 	}
